v0.4.0

+++ kernel/include/stdint.h
@@ -25,7 +25,7 @@ typedef unsigned long size_t;
 typedef long ssize_t;
 typedef unsigned long uintptr_t;
 
-typedef int intmax_t;
+typedef long intmax_t;
 typedef unsigned long uintmax_t;
 
 #endif
